+++ miden-crypto/src/merkle/smt/mod.rs
@@ -29,11 +29,6 @@ pub const SMT_MAX_DEPTH: u8 = 64;
 // SPARSE MERKLE TREE
 // ================================================================================================
 
-/// A map whose keys are not guaranteed to be ordered.
-#[cfg(feature = "smt_hashmaps")]
-type UnorderedMap<K, V> = hashbrown::HashMap<K, V>;
-#[cfg(not(feature = "smt_hashmaps"))]
-type UnorderedMap<K, V> = alloc::collections::BTreeMap<K, V>;
 type InnerNodes = UnorderedMap<NodeIndex, InnerNode>;
 type Leaves<T> = UnorderedMap<u64, T>;
 type NodeMutations = UnorderedMap<NodeIndex, NodeMutation>;