+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [BREAKING] Incremented MSRV to 1.88.
 - [BREAKING] Remove generics from `MerklStore` and remove `KvMap` and `RecordingMap` (#442).
+- Rename `smt_hashmaps` feature to `ext_hashmaps` (#442).
 
 ## 0.15.3 (2025-06-18)